+++ source/gnss/gnssParametrization/gnssParametrizationEarthRotation.h
@@ -20,7 +20,7 @@
 static const char *docstringGnssParametrizationEarthRotation = R"(
 \subsection{EarthRotation}\label{gnssParametrizationType:earthRotation}
 Earth rotation parameters (ERPs) can be estimated by defining
-\config{estimatePole} ($x_p$, $y_p$) and \config{estimateUT1} ($dUT1, LOD$).
+\config{estimatePole} ($x_p$, $y_p\, [mas]$) and \config{estimateUT1} ($dUT1\, [ms], LOD$).
 
 Estimating length of day (LOD) with the sign according to IGS conventions requires a negative
 value in \configClass{parametrizationTemporal:trend:timeStep}{parametrizationTemporalType:trend}.
@@ -30,6 +30,15 @@ Constraints on the defined parameters can be added via
 An example would be to set up \configClass{estimateUT1:constant}{parametrizationTemporalType:constant}
 so the $dUT1$ parameter is included in the normal equation system . Since $dUT1$ cannot be
 determined by GNSS, a hard constraint to its a priori value can then be added.
+
+The \file{parameter names}{parameterName} are
+\begin{itemize}
+\item \verb|earth:polarMotion.xp:<temporal>:<interval>|,
+\item \verb|earth:polarMotion.yp:<temporal>:<interval>|,
+\item \verb|earth:UT1:<temporal>:<interval>|,
+\item \verb|earth:nutation.X:<temporal>:<interval>|,
+\item \verb|earth:nutation.>:<temporal>:<interval>|.
+\end{itemize}
 )";
 #endif