+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Other:            SimulateStarCamera: new optional nadir pointing mode.
 - Other:            SimulateOrbit: Improved temporal numerical precision.
 - Other:            GnssRinexNavigation2OrbitClock: add SBAS support.
+- Other:            Updated documentation to explain parameter names.
 
 # Release 2023-08-25
 - New program:      FileMove.